Extract a Contest.js
